requirements.txt
scapy>=2.5.0
Pillow>=10.0.0
reportlab>=4.0.0
pytest>=7.4.0
pyinstaller>=6.0.0
pandas>=2.0.0
